criação de exception handlers

+++ src/main/java/com/example/ToDoList/controller/Requests.java
@@ -6,10 +6,15 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.persistence.EntityNotFoundException;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
-import java.awt.*;
+import java.net.URI;
 import java.util.List;
 
 @RestController
@@ -28,7 +33,7 @@ public class Requests {
 
     @Operation(summary = "retorna todas as tarefas existentes", method = "GET")
     @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "criação de tarefa concluida"),
+            @ApiResponse(responseCode = "201", description = "criação de tarefa concluida"),
             @ApiResponse(responseCode = "422", description = "Dados de requisição inválida"),
             @ApiResponse(responseCode = "400", description = "Parametros inválidos"),
             @ApiResponse(responseCode = "500", description = "Erro ao realizar busca de dados"),
@@ -41,7 +46,7 @@ public class Requests {
 
     @Operation(summary = "atualiza as tarefas com base no seu id", method = "PUT")
     @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "criação de tarefa concluida"),
+            @ApiResponse(responseCode = "201", description = "criação de tarefa concluida"),
             @ApiResponse(responseCode = "422", description = "Dados de requisição inválida"),
             @ApiResponse(responseCode = "400", description = "Parametros inválidos"),
             @ApiResponse(responseCode = "500", description = "Erro ao realizar busca de dados"),
@@ -55,37 +60,50 @@ public class Requests {
           tarefa.setStatus(novaTarefa.getStatus());
 
           return repository.save(tarefa);
-      }).orElseThrow();
+      })
+            .orElseThrow(()-> new EntityNotFoundException("tarefa não encontrada com id "+id));
     }
 
 
     @Operation(summary = "Adiciona uma tarefa nova", method = "POST")
     @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "criação de tarefa concluida"),
+            @ApiResponse(responseCode = "201", description = "criação de tarefa concluida"),
             @ApiResponse(responseCode = "422", description = "Dados de requisição inválida"),
             @ApiResponse(responseCode = "400", description = "Parametros inválidos"),
             @ApiResponse(responseCode = "500", description = "Erro ao realizar busca de dados"),
 
     })
     @PostMapping
-    public void create(@RequestBody Tarefa tarefa){
-        repository.save(tarefa);
+    public ResponseEntity<Tarefa> create(@RequestBody @Valid Tarefa tarefa){
+
+        Tarefa novaTarefa =repository.save(tarefa);
+
+        URI uri = ServletUriComponentsBuilder.fromCurrentRequest()
+                .path("/{id}")
+                .buildAndExpand(novaTarefa.getId())
+                .toUri();
+
+        return ResponseEntity.created(uri).body(novaTarefa);
     }
 
 
 
     @Operation(summary = "Remove uma tarefa com base no seu id", method = "DELETE")
     @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "criação de tarefa concluida"),
+            @ApiResponse(responseCode = "201", description = "criação de tarefa concluida"),
             @ApiResponse(responseCode = "422", description = "Dados de requisição inválida"),
             @ApiResponse(responseCode = "400", description = "Parametros inválidos"),
             @ApiResponse(responseCode = "500", description = "Erro ao realizar busca de dados"),
 
     })
     @DeleteMapping("/{id}")
-    public void delete(@PathVariable Long id){
-
-        repository.deleteById(id);
+    public ResponseEntity<Tarefa> delete(@PathVariable Long id){
+        if (repository.existsById(id)) {
+            repository.deleteById(id);
+            return ResponseEntity.noContent().build(); // 204 No Content
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
 }
